// src/generated/sources/form26as.ts
export interface Form26AS {
  metadata: DocumentMetadata | undefined;
  profile: TaxpayerProfile | undefined;
  tdsSalary: TdsSalary[];
  tdsOther: TdsOther[];
  tcsDetails: TcsDetail[];
  advanceTax: TaxChallan[];
  selfAssessmentTax: TaxChallan[];
}

export interface TdsSalary {
  tan: string;
  deductorName: string;
  amount: number;
}

export interface TdsOther {
  tan: string;
  deductorName: string;
  section: string;
  amount: number;
}

export interface TcsDetail {
  collectorName: string;
  amount: number;
}

export interface TaxChallan {
  bsrCode: string;
  date: string;
  challanNo: string;
  amount: number;
}
// ...
export const TdsSalary: MessageFns<TdsSalary> = {
  encode(message: TdsSalary, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.tan !== "") {
      writer.uint32(10).string(message.tan);
    }
    if (message.deductorName !== "") {
      writer.uint32(18).string(message.deductorName);
    }
    if (message.amount !== 0) {
      writer.uint32(25).double(message.amount);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TdsSalary {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTdsSalary();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.tan = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.deductorName = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 25) {
            break;
          }

          message.amount = reader.double();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TdsSalary {
    return {
      tan: isSet(object.tan) ? globalThis.String(object.tan) : "",
      deductorName: isSet(object.deductorName)
        ? globalThis.String(object.deductorName)
        : isSet(object.deductor_name)
        ? globalThis.String(object.deductor_name)
        : "",
      amount: isSet(object.amount) ? globalThis.Number(object.amount) : 0,
    };
  },

  toJSON(message: TdsSalary): unknown {
    const obj: any = {};
    if (message.tan !== "") {
      obj.tan = message.tan;
    }
    if (message.deductorName !== "") {
      obj.deductorName = message.deductorName;
    }
    if (message.amount !== 0) {
      obj.amount = message.amount;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TdsSalary>, I>>(base?: I): TdsSalary {
    return TdsSalary.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TdsSalary>, I>>(object: I): TdsSalary {
    const message = createBaseTdsSalary();
    message.tan = object.tan ?? "";
    message.deductorName = object.deductorName ?? "";
    message.amount = object.amount ?? 0;
    return message;
  },
};

function createBaseTdsOther(): TdsOther {
  return { tan: "", deductorName: "", section: "", amount: 0 };
}

export const TdsOther: MessageFns<TdsOther> = {
  encode(message: TdsOther, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.tan !== "") {
      writer.uint32(10).string(message.tan);
    }
    if (message.deductorName !== "") {
      writer.uint32(18).string(message.deductorName);
    }
    if (message.section !== "") {
      writer.uint32(26).string(message.section);
    }
    if (message.amount !== 0) {
      writer.uint32(33).double(message.amount);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TdsOther {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTdsOther();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.tan = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.deductorName = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.section = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 33) {
            break;
          }

          message.amount = reader.double();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TdsOther {
    return {
      tan: isSet(object.tan) ? globalThis.String(object.tan) : "",
      deductorName: isSet(object.deductorName)
        ? globalThis.String(object.deductorName)
        : isSet(object.deductor_name)
        ? globalThis.String(object.deductor_name)
        : "",
      section: isSet(object.section) ? globalThis.String(object.section) : "",
      amount: isSet(object.amount) ? globalThis.Number(object.amount) : 0,
    };
  },

  toJSON(message: TdsOther): unknown {
    const obj: any = {};
    if (message.tan !== "") {
      obj.tan = message.tan;
    }
    if (message.deductorName !== "") {
      obj.deductorName = message.deductorName;
    }
    if (message.section !== "") {
      obj.section = message.section;
    }
    if (message.amount !== 0) {
      obj.amount = message.amount;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TdsOther>, I>>(base?: I): TdsOther {
    return TdsOther.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TdsOther>, I>>(object: I): TdsOther {
    const message = createBaseTdsOther();
    message.tan = object.tan ?? "";
    message.deductorName = object.deductorName ?? "";
    message.section = object.section ?? "";
    message.amount = object.amount ?? 0;
    return message;
  },
};

function createBaseTcsDetail(): TcsDetail {
  return { collectorName: "", amount: 0 };
}

export const TcsDetail: MessageFns<TcsDetail> = {
  encode(message: TcsDetail, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.collectorName !== "") {
      writer.uint32(10).string(message.collectorName);
    }
    if (message.amount !== 0) {
      writer.uint32(17).double(message.amount);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TcsDetail {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTcsDetail();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.collectorName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 17) {
            break;
          }

          message.amount = reader.double();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TcsDetail {
    return {
      collectorName: isSet(object.collectorName)
        ? globalThis.String(object.collectorName)
        : isSet(object.collector_name)
        ? globalThis.String(object.collector_name)
        : "",
      amount: isSet(object.amount) ? globalThis.Number(object.amount) : 0,
    };
  },

  toJSON(message: TcsDetail): unknown {
    const obj: any = {};
    if (message.collectorName !== "") {
      obj.collectorName = message.collectorName;
    }
    if (message.amount !== 0) {
      obj.amount = message.amount;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TcsDetail>, I>>(base?: I): TcsDetail {
    return TcsDetail.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TcsDetail>, I>>(object: I): TcsDetail {
    const message = createBaseTcsDetail();
    message.collectorName = object.collectorName ?? "";
    message.amount = object.amount ?? 0;
    return message;
  },
};

function createBaseTaxChallan(): TaxChallan {
  return { bsrCode: "", date: "", challanNo: "", amount: 0 };
}

export const TaxChallan: MessageFns<TaxChallan> = {
  encode(message: TaxChallan, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.bsrCode !== "") {
      writer.uint32(10).string(message.bsrCode);
    }
    if (message.date !== "") {
      writer.uint32(18).string(message.date);
    }
    if (message.challanNo !== "") {
      writer.uint32(26).string(message.challanNo);
    }
    if (message.amount !== 0) {
      writer.uint32(33).double(message.amount);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TaxChallan {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTaxChallan();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.bsrCode = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.date = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.challanNo = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 33) {
            break;
          }

          message.amount = reader.double();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TaxChallan {
    return {
      bsrCode: isSet(object.bsrCode)
        ? globalThis.String(object.bsrCode)
        : isSet(object.bsr_code)
        ? globalThis.String(object.bsr_code)
        : "",
      date: isSet(object.date) ? globalThis.String(object.date) : "",
      challanNo: isSet(object.challanNo)
        ? globalThis.String(object.challanNo)
        : isSet(object.challan_no)
        ? globalThis.String(object.challan_no)
        : "",
      amount: isSet(object.amount) ? globalThis.Number(object.amount) : 0,
    };
  },

  toJSON(message: TaxChallan): unknown {
    const obj: any = {};
    if (message.bsrCode !== "") {
      obj.bsrCode = message.bsrCode;
    }
    if (message.date !== "") {
      obj.date = message.date;
    }
    if (message.challanNo !== "") {
      obj.challanNo = message.challanNo;
    }
    if (message.amount !== 0) {
      obj.amount = message.amount;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TaxChallan>, I>>(base?: I): TaxChallan {
    return TaxChallan.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TaxChallan>, I>>(object: I): TaxChallan {
    const message = createBaseTaxChallan();
    message.bsrCode = object.bsrCode ?? "";
    message.date = object.date ?? "";
    message.challanNo = object.challanNo ?? "";
    message.amount = object.amount ?? 0;
    return message;
  },
};
// ...
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}